\name{AdmixDataSim2}
\docType{data}
\alias{AdmixDataSim2}
\title{Admixed Population Genotype Data from Simulation 2}
\description{
This data set contains genotypes for a simulated admixed
population. The admixed population consists of 100 individuals scored
for 90 co-dominant markers and 10 dominant markers. Admixed individuals
were simulated by sampling genotypes from parental populations
(\code{p1DataSim2} and \code{p2DataSim2}). The first and second rows of
this matrix data object give population and individual identifications
for each admixed individual. The following rows contain the genotype
data; rows correspond to markers and columns correspond to individuals. 

}
\usage{data(AdmixDataSim2)}
\format{A matrix with 102 rows and 100 columns.}
\source{From simulations conducted by CAB and ZG.}
\seealso{
\code{\link{LociDataSim2}}, \code{\link{p1DataSim2}}, \code{\link{p2DataSim2}}
}

\keyword{datasets}

