\name{fixup.combos.touse}
\alias{fixup.combos.touse}
\title{Fix Combos To Use}
\description{
This is an internal function called by \code{prepare.data}. 

}


\author{
Zachariah Gompert \email{zgompert@uwyo.edu}, C. Alex Buerkle \email{buerkle@uwyo.edu}
}  

\seealso{
\code{\link{test.combinations}}, \code{\link{prepare.data}}
} 
 