\name{LociDataSim2}
\docType{data}
\alias{LociDataSim2}
\title{Marker Information for Simulation 2}
\description{
This data object provides the marker information associated with
\code{AdmixDataSim2}. Rows correspond to markers (100 markers total). The first
column gives the locus name and the second column gives the locus type;
co-dominant (C) for the first 90 markers, dominant (D) for the last 10
markers. Map information is not provided for these markers.
}
\usage{data(LociDataSim2)}
\format{A matrix with 100 rows and 2 columns.}
\source{From simulations conducted by CAB and ZG.}
\seealso{
\code{\link{AdmixDataSim2}}, \code{\link{p1DataSim2}}, \code{\link{p2DataSim2}}
}

\keyword{datasets}

