\name{p1D}
\docType{data}
\alias{p1D}
\title{Parental Population 1 Dominant Genotype Data from Simulation 1}
\description{
This data set contains genotype data for parental population 1
associated with \code{AdmixDom}. This data set consists of 110 dominant
markers scored for 10 individuals. All individuals were given a genotype
of 1, meaning they possessed the dominant allele at each marker.
}
\usage{data(p1D)}
\format{A matrix with 110 rows and 10 columns.}
\source{From simulations conducted by CAB and ZG.}
\seealso{
\code{\link{AdmixDom}}
}

\keyword{datasets}

