\name{p2D}
\docType{data}
\alias{p2D}
\title{Parental Population 2 Dominant Genotype Data from Simulation 1}
\description{
This data set contains genotype data for parental population 2
associated with \code{AdmixDom}. This data set consists of 110 dominant
markers scored for 10 individuals. All individuals were given a genotype
of 0, meaning that they are homozygous for the recessive allele at each marker.
}
\usage{data(p2D)}
\format{A matrix with 110 rows and 10 columns.}
\source{From simulations conducted by CAB and ZG.}
\seealso{
\code{\link{AdmixDom}}
}

\keyword{datasets}

