\name{p2DataSim2}
\docType{data}
\alias{p2DataSim2}
\title{Parental Population 2 Genotype Data from Simulation 2}
\description{
This data set contains genotype data for parental population 2
associated with \code{AdmixDataSim2}. This data set consists of 90
co-dominant and 10 dominant
markers scored for 50 individuals. The 224 and 230 alleles occurs at high
frequency for the 90 co-dominant markers in this population, as does the
0 (recessive) allele for the 10 dominant markers.
}
\usage{data(p2DataSim2)}
\format{A matrix with 100 rows and 50 columns.}
\source{From simulations conducted by CAB and ZG.}
\seealso{
\code{\link{AdmixDataSim2}}
}

\keyword{datasets}

