\name{test.combinations}
\alias{test.combinations}
\title{Test Combinations}
\description{
This is an internal function called by \code{prepare.data}. 

}


\author{
Zachariah Gompert \email{zgompert@uwyo.edu}, C. Alex Buerkle \email{buerkle@uwyo.edu}
}  

\seealso{
\code{\link{delta}}, \code{\link{prepare.data}}
} 
 